Dependent tasks
This change represents a change in my thinking
and a simplification of the API: I was originally
thinking that all tasks were by default forced to
be serial, and that the only way to opt into
concurrency was to start defining concurrency
constraints

But since adding the concept of "dependent" tasks,
I realized this concept could be extended to elegantly
allow the user to opt into non-concurrency in an
extremely simple manner, using task() with a subtask
defined. I'll write some docs on this soon.