Expose properties on Task that show fulfill/reject values
If you perform a task and it runs to completion,
it'd be nice if the value of the completed task
instance were exposed on the task object so that
it could be displayed in the template. Without
this feature, you have to .set some property
to expose it to the template.

Rather than just exposing the value/error directly
on the task object (which would lead to a combinatoric
explosion of property names to cover all the various
use cases), this commit exposes a handful of new
properties on Task, each of which points to a
TaskInstance (a single execution of a task). In
addition, the completion/cancelation/error state
is now exposed under the TaskInstance properties
of `.value` and `.error`.

So if you want to display the last successful
return value of a task, you would use

    {{myTask.lastSuccessful.value}}

where `.lastSuccessful` is one of the newly added
properties of Task which points to a TaskInstance,
and `.value` is the value returned from the
most recent task instance that ran to completion.

Depending on what you need for your UI, you can
use a different combination of task instance and
.value/.error (e.g. sometimes you want to keep
displaying the last success even if it's followed
by failed/canceled attempts, other times you want
to clear it out right away).

The newly added Task properties are:

- lastPerformed
  - the taskInstance from the last perform, regardless of
    whether it was dropped/enqueued/immediate
- lastStarted
  - the taskInstance the most recently started
    - .drop()ped tasks won't show up here
    - .enqueue()d tasks only show up here when they start
- lastSuccessful
  - the task instance that ran to completion and returned a value
  - corrollary: lastSuccessful.error can never be a thing
- lastComplete
  - the task instance that most recently completed/errored/cancelled
- lastErrored
  - the task instance that most recently errored (not cancelled)
- lastCanceled (.value, .error)
  - the task instance that most recently cancelled (not error)
- lastIncomplete
  - the task instance that errored or canceled

/cc @lolmaus