Support null/undefined values in cancelable promise helpers
Similarly to how it's very nice to be able to pass a mix of promises and simple values to Ember.RSVP.{hash,all}, it's nice to be able to pass a mix of task results and simple values to ember-concurrency's hash() and all() helpers. However, if one of these values were null or undefined, the cancellation code would throw an error. This fixes that.