Remove any assumptions about how you track state
The numPerformedInc is a tricky property because,
it represents a delta/increment, as opposed to a
full count (such as numRunning/numQueued, which
can be fully re-calculated every time). Previously,
we were converting it to a non-delta performCount
within the SchedulerRefresh, but I want to remove
any knowledge of how the state is stored on the
task itself from the ScheduleRefresh classes.