Add ember-source@3.8 + decorators polyfill to CI matrix.
Support for using `Ember.computed` as a decorator was added in Ember
3.10 (which we are testing against already), however
`ember-decorators-polyfill` can also be used to add that support to
older Ember versions.

Since ember-source@3.8 is the current LTS it is a common target for
folks that are looking for overall stability. If those folks also want
to author in the newer style they would use the
ember-decorators-polyfill.