Add some brief documentation on use with Octane/decorators
Raised in issue #271, there is currently no official documentation
on using ember-concurrency with native classes and other Octane-related
use-cases.

This adds a brief section to "Task Function Syntax" in the docs that
shows an example of use with a Glimmer component and mentions the
ember-concurrency-decorators addon as a more concise approach to the
default decorators via the task() property.