Remove generics default values
We don't really have a case where these default values were helpful,
but having them around seems to make the inferencer work "less hard"
in some cases – where it would fallback to the default values when
it could otherwise infer something more useful when the defaults are
not provided.

Also uses `infer` explicitly in the `task` functions as it seems to
give better results.