Fix build
Some of the new development dependencies does not support Node 8.
These packages doesn't really affect runtime behavior, and are
only used in the examples (which are checked by `tsc` to ensure
they are correct).

As a workaround, we are making most jobs on CI run on Node 10, but
keep a matrix row for Node 8 that does not try to install those
dependencies or try to typecheck the code.