Deprecate direct usage of task with action helper and remove tests for it on 3.20+
This feature unfortunately depends on private APIs that will be
removed in Ember 3.25. Unforunately, the mechanism was already
removed in Ember 3.20, making it impossible to support this for
releases at least until recent canaries. However, the feature is
rarely used, only documented in the changelog, and frequently
breaks whenever the private constant it depends on moves modules
internally throughout new versions of Ember.

It will continue to remain available for Ember < 3.20, but will
not be available in Ember-Concurrency 2.0, and will show a
deprecation warning from 1.3.0 forward.