Fix encapsulated state render updates
This one is a bit in-the-weeds, but it's effectively
a bugfix plus a workaround for < 3.16. We're now wrapping
TaskInstance internally and externally for EncapsulatedTask
launched instances. This is so that internal state updates
work. However, for Ember < 3.16, we needed to go a step
further and add some additional handlers to the Proxy setup
in order to ensure derived state updates get rendered
correctly, due to the way the state was stored in those
versions.