Refactor Task and TaskGroup decorator and properties
Introduces a TaskFactory and TaskGroupFactory that can be made
de-coupled from the Ember object model. Avoids reliance on computed
property internals, while still providing TaskProperty compatibility.

This will set the stage for potential future public APIs for making
tasks more extensible by hooking into the creation of them, while also
reducing the tediousness of instantiating Tasks and TaskGroup directly
(which is not yet public API either, but may be in the future.)