Restore access to host context from encapsulated tasks
It was inadvertently removed when re-implementing encapsulated
tasks with Proxy. However, it's a useful feature for implementing
sort-of "higher-order tasks" and since there's no viable replacement,
this PR reintroduces it. It is, however, considered intimate,
undocumented API that will not be covered under SemVer guarantees.

Because it can be read from `this`, it can be accidentally overwritten
by user code. A future solution might be to provide some sort of
`getHostContext()` yieldable to provide the value within task execution
and avoid the need to access it directly.

For now, restoring it will allow folks who depended on such undocumented
behavior to upgrade to 2.0.0.