Fix CI build
Something (probably a release of new embroider packages) caused most of the floating dependencies scenarios to start failing in all branches. It looks like it was some kind of mismatch between the version of `@embroider/macros` used by `@ember/render-modifiers` (a dependency of `ember-prism`) and whatever version of `@embroider/shared-internals` it was finding in the `node_modules` tree.

So this commit updates us to the latest `@embroider/macros`, which allows us to dedupe `@embroider/macros` and `@embroider/shared-internals`, which appears to fix the issue.